(PC-35100) feat(ui): use GenericInfoPageWhite in ConfirmChangeEmail
