+++ src/features/profile/pages/ConfirmChangeEmail/ConfirmChangeEmail.tsx
@@ -1,20 +1,81 @@
-import React from 'react'
+import { useNavigation, useRoute } from '@react-navigation/native'
+import React, { useCallback } from 'react'
 
-import { ConfirmChangeEmailContent } from 'features/profile/pages/ConfirmChangeEmail/ConfirmChangeEmailContent'
-import { GenericInfoPageWhiteLegacy } from 'ui/pages/GenericInfoPageWhiteLegacy'
+import { ApiError } from 'api/ApiError'
+import { AccountState } from 'api/gen'
+import { useLoginRoutine } from 'features/auth/helpers/useLoginRoutine'
+import { navigateToHomeConfig } from 'features/navigation/helpers/navigateToHome'
+import { UseNavigationType, UseRouteType } from 'features/navigation/RootNavigator/types'
+import { useConfirmChangeEmailMutationV2 } from 'features/profile/helpers/useConfirmChangeEmailMutationV2'
+import { isTimestampExpired } from 'libs/dates'
+import { SNACK_BAR_TIME_OUT, useSnackBarContext } from 'ui/components/snackBar/SnackBarContext'
+import { GenericInfoPageWhite } from 'ui/pages/GenericInfoPageWhite'
 import { BicolorPhonePending } from 'ui/svg/icons/BicolorPhonePending'
-import { Spacer, TypoDS } from 'ui/theme'
+import { Invalidate } from 'ui/svg/icons/Invalidate'
 
 export function ConfirmChangeEmail() {
+  const { replace, reset } = useNavigation<UseNavigationType>()
+  const { params } = useRoute<UseRouteType<'ConfirmChangeEmail'>>()
+  const loginRoutine = useLoginRoutine()
+  const { showErrorSnackBar } = useSnackBarContext()
+  const { mutate, isLoading } = useConfirmChangeEmailMutationV2({
+    onSuccess: async ({
+      accessToken,
+      refreshToken,
+      newEmailSelectionToken,
+      resetPasswordToken,
+    }) => {
+      await loginRoutine(
+        { accessToken, refreshToken, accountState: AccountState.ACTIVE },
+        'fromConfirmChangeEmail'
+      )
+
+      if (resetPasswordToken) {
+        replace('ChangeEmailSetPassword', {
+          token: resetPasswordToken,
+          emailSelectionToken: newEmailSelectionToken,
+        })
+      } else {
+        replace('NewEmailSelection', { token: newEmailSelectionToken })
+      }
+    },
+    onError: (error) => {
+      if (error instanceof ApiError && error.statusCode === 401) {
+        reset({ index: 0, routes: [{ name: 'ChangeEmailExpiredLink' }] })
+        return
+      }
+      showErrorSnackBar({
+        message: 'Désolé, une erreur technique s’est produite. Veuillez réessayer plus tard.',
+        timeout: SNACK_BAR_TIME_OUT,
+      })
+    },
+  })
+
+  const onConfirmEmail = useCallback(
+    () => mutate({ token: params?.token }),
+    [params?.token, mutate]
+  )
+
+  if (isTimestampExpired(params.expiration_timestamp)) {
+    reset({ index: 0, routes: [{ name: 'ChangeEmailExpiredLink' }] })
+    return null
+  }
+
   return (
-    <GenericInfoPageWhiteLegacy
-      icon={BicolorPhonePending}
-      titleComponent={TypoDS.Title3}
+    <GenericInfoPageWhite
+      illustration={BicolorPhonePending}
       title="Confirmes-tu la demande de changement d’e-mail&nbsp;?"
-      separateIconFromTitle={false}
-      mobileBottomFlex={0.3}>
-      <Spacer.Column numberOfSpaces={40} />
-      <ConfirmChangeEmailContent />
-    </GenericInfoPageWhiteLegacy>
+      buttonPrimary={{
+        wording: 'Confirmer la demande',
+        onPress: onConfirmEmail,
+        disabled: isLoading,
+      }}
+      buttonTertiary={{
+        wording: 'Annuler',
+        navigateTo: navigateToHomeConfig,
+        icon: Invalidate,
+        disabled: isLoading,
+      }}
+    />
   )
 }
